+++ python/aristotle-mdr-graphql/aristotle_mdr_graphql/tests/tests.py
@@ -4,6 +4,7 @@ from aristotle_mdr.tests import utils
 from aristotle_mdr import models as mdr_models
 from aristotle_dse import models as dse_models
 from comet import models as comet_models
+from graphene.test import Client as QLClient
 
 import json
 import datetime
@@ -14,6 +15,7 @@ class BaseGraphqlTestCase(utils.LoggedInViewPages):
 
         super().setUp()
         self.client = Client()
+
         self.apiurl = reverse('aristotle_graphql:graphql_api')
 
         self.dec = mdr_models.DataElementConcept.objects.create(
@@ -72,8 +74,16 @@ class GraphqlFunctionalTests(BaseGraphqlTestCase, TestCase):
         self.login_editor()
 
         response = self.client.get(self.apiurl, HTTP_ACCEPT='text/html')
+        self.assertRedirects(response, reverse('aristotle_graphql:graphql_explorer'))
+
+        response = self.client.get(self.apiurl+"?noexplorer", HTTP_ACCEPT='text/html')
         self.assertEqual(response.status_code, 200)
         self.assertTemplateUsed('graphene/graphiql.html')
+
+
+        response = self.client.get(reverse('aristotle_graphql:graphql_explorer'))
+        self.assertEqual(response.status_code, 200)
+        self.assertTemplateUsed('aristotle_mdr_graphql/explorer.html')
 
     def test_query_by_uuid(self):
 
@@ -108,29 +118,29 @@ class GraphqlFunctionalTests(BaseGraphqlTestCase, TestCase):
     def test_dse_query(self):
 
         self.login_editor()
-        dse_models.Dataset.objects.create(
-            name='Test Dataset',
+        dse_models.DataSetSpecification.objects.create(
+            name='Test DSS',
             definition='Test Defn',
             workgroup=self.wg1
         )
 
-        response_json = self.post_query('{ datasets { edges { node { name } } } }')
-        edges = response_json['data']['datasets']['edges']
+        response_json = self.post_query('{ datasetSpecifications { edges { node { name } } } }')
+        edges = response_json['data']['datasetSpecifications']['edges']
 
         self.assertEqual(len(edges), 1)
-        self.assertEqual(edges[0]['node']['name'], 'Test Dataset')
+        self.assertEqual(edges[0]['node']['name'], 'Test DSS')
 
     def test_comet_query(self):
 
         self.login_editor()
-        comet_models.IndicatorSet.objects.create(
+        comet_models.Indicator.objects.create(
             name='Test Indicator Set',
             definition='Test Defn',
             workgroup=self.wg1
         )
 
-        response_json = self.post_query('{ indicatorSets { edges { node { name } } } }')
-        edges = response_json['data']['indicatorSets']['edges']
+        response_json = self.post_query('{ indicators { edges { node { name } } } }')
+        edges = response_json['data']['indicators']['edges']
 
         self.assertEqual(len(edges), 1)
         self.assertEqual(edges[0]['node']['name'], 'Test Indicator Set')
@@ -160,32 +170,41 @@ class GraphqlFunctionalTests(BaseGraphqlTestCase, TestCase):
     def test_query_related_m2m(self):
         # Test a query on an items many to many relation
 
-        rr = mdr_models.ReviewRequest.objects.create(
-            requester=self.editor,
-            registration_authority=self.ra,
-            status=0,
-            state=0,
-            registration_date=datetime.date.today(),
-            cascade_registration=0
+        ded = mdr_models.DataElementDerivation.objects.create(
+            submitter=self.editor,
+            name="My Calculation"
         )
 
-        rr.concepts.add(self.de)
-        rr.concepts.add(self.dec)
-        rr.concepts.add(self.vd)
+        ded.inputs.add(self.de)
+
+        self.assertTrue(ded.can_view(self.editor))
+        self.assertTrue(self.de.can_view(self.editor))
 
         self.login_editor()
 
-        json_response = self.post_query('{ reviewRequests { edges { node { concepts { edges { node { name } } } } } } }')
-        edges = json_response['data']['reviewRequests']['edges']
+        query = '{ dataElementDerivations { edges { node { name inputs { edges { node { name } } } } } } }'
+        json_response = self.post_query(query)
+        edges = json_response['data']['dataElementDerivations']['edges']
         self.assertEqual(len(edges), 1)
 
-        concept_edges = edges[0]['node']['concepts']['edges']
-        self.assertEqual(len(concept_edges), 3)
+        concept_edges = edges[0]['node']['inputs']['edges']
+        self.assertEqual(len(concept_edges), 1)
 
-        item_names = [self.de.name, self.dec.name, self.vd.name]
+        item_names = [self.de.name]
 
         for item in concept_edges:
             self.assertTrue(item['node']['name'] in item_names)
+
+        # Test accessing an item user doesnt have permission to view through a many to many relation
+        self.de.workgroup = None
+        self.de.save()
+        self.de = self.de.__class__.objects.get(pk=self.de.pk)
+        self.assertFalse(self.de.can_view(self.editor))
+
+        json_response = self.post_query(query)
+        edges = json_response['data']['dataElementDerivations']['edges']
+        concept_edges = edges[0]['node']['inputs']['edges']
+        self.assertEqual(len(concept_edges), 0)
 
     def test_query_table_inheritance(self):
         # Test a query of a table inheritance property (from metadata to dataelement)
@@ -203,6 +222,7 @@ class GraphqlFunctionalTests(BaseGraphqlTestCase, TestCase):
         self.assertEqual(len(edges), 1)
         self.assertEqual(edges[0]['node']['name'], self.dec.name)
         self.assertEqual(edges[0]['node']['dataelement'], None)
+
 
 class GraphqlPermissionsTests(BaseGraphqlTestCase, TestCase):
 
@@ -244,6 +264,8 @@ class GraphqlPermissionsTests(BaseGraphqlTestCase, TestCase):
         self.vd._is_public = True
         self.vd.save()
 
+        self.assertTrue(self.vd.can_view(self.editor))
+
         json_response = self.post_query('{ dataElements { edges { node { name } } } }')
         self.assertEqual(len(json_response['data']['dataElements']['edges']), 0)
 
@@ -261,6 +283,7 @@ class GraphqlPermissionsTests(BaseGraphqlTestCase, TestCase):
         self.vd.save()
 
         self.login_editor()
+        self.assertFalse(self.vd.can_view(self.editor))
 
         json_response = self.post_query('{ dataElements { edges { node { name dataElementConcept { name } valueDomain { name } } } } }')
         edges = json_response['data']['dataElements']['edges']
@@ -283,64 +306,35 @@ class GraphqlPermissionsTests(BaseGraphqlTestCase, TestCase):
         self.assertEqual(edges[0]['node']['name'], 'Test Value Domain')
         self.assertEqual(len(edges[0]['node']['dataelementSet']['edges']), 0)
 
-    def test_query_not_allowed_m2m(self):
-        # Test accessing an item user doesnt have permission to view through a many to many relation
+    # Filtering out RRs for now.
+    # When we can perform actions against RRs, then we'll bring them back
+    # def test_reviewrequest_query_perms(self):
 
-        self.vd.workgroup = self.wg2
-        self.vd.save()
+    #     allowed_rr = mdr_models.ReviewRequest.objects.create(
+    #         requester=self.editor,
+    #         registration_authority=self.ra,
+    #         status=0,
+    #         state=1,
+    #         registration_date=datetime.date.today(),
+    #         cascade_registration=0
+    #     )
 
-        rr = mdr_models.ReviewRequest.objects.create(
-            requester=self.editor,
-            registration_authority=self.ra,
-            status=0,
-            state=0,
-            registration_date=datetime.date.today(),
-            cascade_registration=0
-        )
+    #     disallowed_rr = mdr_models.ReviewRequest.objects.create(
+    #         requester=self.viewer,
+    #         registration_authority=self.ra,
+    #         status=0,
+    #         state=0,
+    #         registration_date=datetime.date.today(),
+    #         cascade_registration=0
+    #     )
 
-        rr.concepts.add(self.de)
-        rr.concepts.add(self.dec)
-        rr.concepts.add(self.vd)
+    #     self.login_editor()
 
-        self.login_editor()
+    #     json_response = self.post_query('{ reviewRequests { edges { node { id state } } } }')
+    #     edges = json_response['data']['reviewRequests']['edges']
 
-        json_response = self.post_query('{ reviewRequests { edges { node { concepts { edges { node { name } } } } } } }')
-        edges = json_response['data']['reviewRequests']['edges']
-        self.assertEqual(len(edges), 1)
-
-        concept_edges = edges[0]['node']['concepts']['edges']
-        self.assertEqual(len(concept_edges), 2)
-
-        for item in concept_edges:
-            self.assertNotEqual(item['node']['name'], 'Test Value Domain')
-
-    def test_reviewrequest_query_perms(self):
-
-        allowed_rr = mdr_models.ReviewRequest.objects.create(
-            requester=self.editor,
-            registration_authority=self.ra,
-            status=0,
-            state=1,
-            registration_date=datetime.date.today(),
-            cascade_registration=0
-        )
-
-        disallowed_rr = mdr_models.ReviewRequest.objects.create(
-            requester=self.viewer,
-            registration_authority=self.ra,
-            status=0,
-            state=0,
-            registration_date=datetime.date.today(),
-            cascade_registration=0
-        )
-
-        self.login_editor()
-
-        json_response = self.post_query('{ reviewRequests { edges { node { id state } } } }')
-        edges = json_response['data']['reviewRequests']['edges']
-
-        self.assertEqual(len(edges), 1)
-        self.assertEqual(edges[0]['node']['state'], 'A_1')
+    #     self.assertEqual(len(edges), 1)
+    #     self.assertEqual(edges[0]['node']['state'], 'A_1')
 
     def test_query_non_registered_item(self):
         # Test requesting an object without a defined node e.g. User